+++ baseParser.go
@@ -1,32 +1,22 @@
 package inigo
 
+import "strings"
+
 // @Date：   2018/8/19 0019 14:25
 // @Author:  Joshua Conero
 // @Name:    基本 go 解析器
+const (
+	baseCommentReg = "^#|;"               // 注释符号
+	baseSectionReg = "^\\[[^\\[\\]]+\\]$" // 节正则
+
+	baseEqualToken = "="      // 等于符号
+	baseSecRegPref = "__sec_" // 节前缀
+)
 
 type BaseParser struct {
-	Data  map[string]interface{}
-	valid bool
-}
-
-// data 数据检测
-func (p *BaseParser) GetData() map[string]interface{} {
-	if p.Data == nil {
-		p.Data = map[string]interface{}{}
-	}
-	return p.Data
-}
-
-func (p *BaseParser) Get(key string) (bool, interface{}) {
-	data := p.GetData()
-	value, has := data[key]
-	return has, value
-}
-
-func (p *BaseParser) HasKey(key string) bool {
-	data := p.GetData()
-	_, has := data[key]
-	return has
+	valid   bool
+	section []string
+	container
 }
 
 func (p *BaseParser) Raw(key string) string {
@@ -34,25 +24,36 @@ func (p *BaseParser) Raw(key string) string {
 	return raw
 }
 
-func (p *BaseParser) Value(params ...interface{}) interface{} {
-	// key, nil, def
-	if len(params) > 2 {
-		if has, value := p.Get(params[0].(string)); has {
-			return value
-		}
-		return params[2]
-	} else if len(params) > 1 {
-		p.Set(params[0].(string), params[1])
-	} else if len(params) == 1 {
-		if has, value := p.Get(params[0].(string)); has {
-			return value
-		}
-	}
-	return nil
+func (p *BaseParser) GetAllSection() []string {
+	return p.section
 }
 
-func (p *BaseParser) GetAllSection() []string {
-	return nil
+func (p *BaseParser) Section(params ...interface{}) interface{}{
+	var value interface{}
+	var section, key string
+
+	if nil == params{
+		value = nil
+	}else if len(params) == 1{
+		format := params[0].(string)
+		if idx := strings.Index(format, "."); idx > -1{
+			section = format[0: idx]
+			key = format[idx+1:]
+		}
+	}else if len(params) > -1{
+		section = params[0].(string)
+		key = params[1].(string)
+	}
+
+	if section != "" && key != ""{
+		if data, hasSection := p.Data[baseSecRegPref+section]; hasSection{
+			dd := data.(map[interface{}]interface{})
+			if v, hasKey := dd[key]; hasKey{
+				value = v
+			}
+		}
+	}
+	return value
 }
 
 func (p *BaseParser) Set(key string, value interface{}) Parser {
@@ -66,11 +67,13 @@ func (p *BaseParser) IsValid() bool {
 }
 
 func (p *BaseParser) OpenFile(filename string) Parser {
+	reader := &baseFileParse{}
+	reader.read(filename)
+	p.Data = reader.GetData()
 	return p
 }
 
 func (p *BaseParser) ReadStr(content string) Parser {
-
 	return p
 }
 
@@ -82,11 +85,11 @@ type BaseStrParse struct {
 	line int
 }
 
-func (p *BaseStrParse) Line() int{
+func (p *BaseStrParse) Line() int {
 	return p.line
 }
 
-func (p *BaseStrParse) GetData() map[interface{}]interface{}{
+func (p *BaseStrParse) GetData() map[interface{}]interface{} {
 	return p.data
 }